requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
apricot-select==0.6.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.5
biopython==1.85
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
fair-esm @ file:///home2/Cj/workspace/hemopi2/esm
fcd-torch==1.0.7
filelock==3.18.0
flow_matching==1.0.10
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.5.1
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
hf-xet==1.1.5
huggingface-hub==0.33.0
idna==3.10
imbalanced-learn==0.13.0
ipykernel==6.30.1
ipython==8.37.0
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
Levenshtein==0.27.1
lightning-utilities==0.15.1
llvmlite==0.44.0
lxml==6.0.1
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
modlamp==4.3.2
mpmath==1.3.0
multidict==6.6.3
mysql-connector-python==9.4.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
nose==1.3.7
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
pandas==2.3.0
parso==0.8.4
pastel==0.2.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
poethepoet==0.10.0
pomegranate==1.1.2
POT==0.9.5
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pynndescent==0.5.13
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-Levenshtein==0.27.1
pytorch-lightning==2.5.2
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.1
RapidFuzz==3.13.0
rdkit==2025.3.3
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
sentencepiece==0.2.0
sentry-sdk==2.32.0
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
soupsieve==2.8
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.2
tomlkit==0.13.3
torch==2.7.1
torchcde==0.2.5
torchcfm==1.0.7
torchdiffeq==0.2.5
torchdyn==1.0.6
torchmetrics==1.8.0
torchsde==0.2.6
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
trampoline==0.1.2
transformers==4.52.4
triton==3.3.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
umap-learn==0.5.8
urllib3==2.5.0
wandb==0.21.0
wcwidth==0.2.13
widgetsnbextension==4.0.14
yarl==1.20.1
